+++ upgrade_readme_usage.py
@@ -129,6 +129,6 @@ if __name__ == "__main__":
 
     if new_string != old_string:
         logging.basicConfig(level=logging.DEBUG)
-        with pymince.logging.timed_block("upgrade_readme_usage"):
-            with open(readme_path, mode="w", encoding="utf-8") as f:
-                f.write(new_string)
+        timed_block = pymince.logging.timed_block("upgrade_readme_usage")
+        with timed_block, open(readme_path, mode="w", encoding="utf-8"):
+            f.write(new_string)