+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## In Progress ##
 
+## v2.2.0 (2023-01-12) ##
+
 - **Fixed:** dictionary.find_leaf_value. Infinite recursion for args: *key="a", dictionary={"a": "b", "b": "a"}*
 - **Added:** New tests. Better coverage
 - **Added:** text.are_anagram, text.fullstr.are_anagram